Create 18_Maximum_Sum_Circular_Subarray.cpp
You should know kadane algorithm.
As we know the kadane algorithm is to find the max subarray, So with that help, we can find the max subarray and we will find the min subarray with the same concept. We will also sum up all elements of the array. We can check if the sum is equal to the min subarray-sum then we will return max subarray-sum ELSE we will return max(max_subArray_sum,sum-mini).